Update tokens.json
Okay, so I've replaced the link in the tokens.json file, but still trying to figure out how to add the logo on a local repo. I've created one, uploaded the logo, which I believe I changed the URL to the one you've provided but still trying to figure it out.