Gold ART Token icon added
I am adding the icon for Gold ART Token as a favor for @Huddo606 on Telegram. Info for the token can be found here: https://artbyhuddo.com/gold-art-token/
Any question regarding the specifics of the token and liquidity should go to him. I can be reached on Telegram as @TongvaPower.